isolate the quantization modeling code
