+++ aphrodite/quantization/aqlm.py
@@ -9,9 +9,9 @@ import torch
 from torch.nn.parameter import Parameter
 import torch.nn.functional as F
 
-from aphrodite._C import ops
+from aphrodite._quant_C import quant_ops as ops
 from aphrodite.modeling.layers.linear import LinearMethodBase, set_weight_attrs
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
