+++ aphrodite/quantization/eetq.py
@@ -5,7 +5,7 @@ import torch
 from torch.nn.parameter import Parameter
 
 from aphrodite.modeling.layers.linear import LinearMethodBase, set_weight_attrs
-from aphrodite.modeling.layers.quantization.base_config import \
+from aphrodite.quantization.base_config import \
     QuantizationConfig
 
 HAS_EETQ = False
@@ -72,7 +72,7 @@ class EETQConfig(QuantizationConfig):
 
     def support_fused_moe(self) -> bool:
         return False
-    
+
     def rope_style(self) -> Optional[bool]:
         return None
