+++ aphrodite/quantization/exl2.py
@@ -2,10 +2,10 @@ from typing import Any, Dict, List, Optional
 
 import torch
 
-from aphrodite._C import ops
+from aphrodite._quant_C import quant_ops as ops
 from aphrodite.modeling.layers.linear import (LinearMethodBase,
                                               set_weight_attrs)
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
