+++ aphrodite/quantization/quip.py
@@ -3,12 +3,12 @@ from typing import Any, Dict, List, Optional
 import torch
 from torch.nn.parameter import Parameter
 
-from aphrodite._C import ops
+from aphrodite._quant_C import quant_ops as ops
 from aphrodite.modeling.layers.linear import (LinearMethodBase,
                                               set_weight_attrs)
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
-from aphrodite.modeling.layers.quantization.quip_utils import (
+from aphrodite.quantization.quip_utils import (
     get_packed_abs_grid,
     get_hadK,
     matmul_hadUt_cuda,
